day8/demo_mvp/pyproject.toml
[project]
name = "demo-mvp-rye-project"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "lazycoder1", email = "gautamgsabhahit@gmail.com" }
]
dependencies = [
    "google-generativeai>=0.8.5",
    "pyautogui>=0.9.54",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.0",
    "playwright>=1.52.0",
    "jupyterlab>=4.4.3",
    "requests>=2.32.3",
    "pillow>=11.2.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.12",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/demo_mvp_rye_project"]

[project.scripts]
demo = "src.main:main"
